+++ next/src/routes/_layout/questionnaire.$questionnaireId/_layout-q/codes-list/$codesListId.tsx
@@ -1,6 +1,7 @@
 import { useSuspenseQuery } from '@tanstack/react-query'
 import { createFileRoute } from '@tanstack/react-router'
 
+import { codesListFromIdQueryOptions } from '@/api/codesLists'
 import { questionnaireQueryOptions } from '@/api/questionnaires'
 import { variablesQueryOptions } from '@/api/variables'
 import EditCodesList from '@/components/codesLists/edit/EditCodesList'
@@ -24,6 +25,9 @@ export const Route = createFileRoute(
     params: { codesListId, questionnaireId },
   }) => {
     queryClient.ensureQueryData(questionnaireQueryOptions(questionnaireId))
+    queryClient.ensureQueryData(
+      codesListFromIdQueryOptions(questionnaireId, codesListId),
+    )
     queryClient.ensureQueryData(variablesQueryOptions(questionnaireId))
     return { crumb: t('crumb.codesList', { id: codesListId }) }
   },
@@ -32,18 +36,14 @@ export const Route = createFileRoute(
 function RouteComponent() {
   const { questionnaireId, codesListId } = Route.useParams()
   const {
-    data: { codesLists, formulasLanguage },
+    data: { formulasLanguage },
   } = useSuspenseQuery(questionnaireQueryOptions(questionnaireId))
+  const { data: codesList } = useSuspenseQuery(
+    codesListFromIdQueryOptions(questionnaireId, codesListId),
+  )
   const { data: variables } = useSuspenseQuery(
     variablesQueryOptions(questionnaireId),
   )
-
-  let codesList
-  if (codesLists) {
-    for (const element of codesLists) {
-      if (element.id === codesListId) codesList = element
-    }
-  }
 
   return (
     <EditCodesListLayout codesList={codesList}>