+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add a warning modal when editing a code list used by questions, warning the user that the collected variables will be regenerated.
+
 ## [3.8.2](https://github.com/InseeFr/Pogues/releases/tag/3.8.2) - 2026-09-07
 
 ### Fixed